Updated naming and removed extra files
