--- Assets/NiclsInterface/DataPointNicls.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d62268e33a9ae4f7bab25d517a0d50a6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
