--- Assets/NiclsInterface/NiclsInterface2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 90bb77b529fc646a49ae78769d1108df
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
