--- Assets/NiclsInterface/NiclsInterface3.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e42f83d36f1264485aa381adb0e847a2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
